felicia/requirements.txt
Django==2.2.28
djangorestframework==3.12.4
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.4.0